new change in db and add some new page

+++ routes/web.php
@@ -1,9 +1,12 @@
 <?php
 
+use App\Http\Controllers\AdminController;
 use App\Http\Controllers\PagesContreller;
 use App\Http\Controllers\PostsController;
 use App\Http\Controllers\ProfileController;
 use App\Http\Controllers\CommentController;
+use App\Models\User;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Route;
 use Symfony\Component\HttpKernel\Profiler\Profile;
 
@@ -23,6 +26,8 @@ use Symfony\Component\HttpKernel\Profiler\Profile;
 Route::resource('/blog',PostsController::class);
 Route::resource('/profile',ProfileController::class);
 
+Route::resource('/admin',AdminController::class);
+
 
 Route::get('/',[PagesContreller::class, 'index']);
 Route::get('/search',[PagesContreller::class, 'search']);
@@ -36,9 +41,14 @@ function($id){
     if($like){
     DB::table('posts')->where('id',$id)->decrement('count');}
 
-    DB::table('likes')->where('post_id',$id)->delete();
+    DB::table('likes')->where('post_id',$id)->where('user_id',Auth::user()->id)->delete();
     return back()->with('dislike' ,'You dont like this post');
+
+
 });
+
+Route::get('/admin/accepted/{postId}',[AdminController::class, 'acceptPost']);
+
 
 
 Route::post('/blog/comments',[CommentController::class, 'store']);
